feat: dao support multi-tenant
--story=120227834
