--- src/storage/dal/mongo/local/local.go
@@ -1,2 +1,0 @@
-// Package local TODO
-package local
