+++ src/storage/dal/mongo/local/mongo.go
@@ -10,13 +10,13 @@
  * limitations under the License.
  */
 
+// Package local is the local mongodb dao implementation
 package local
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	"reflect"
 	"strconv"
 	"strings"
 	"time"
@@ -26,29 +26,30 @@ import (
 	"configcenter/src/common/blog"
 	"configcenter/src/common/json"
 	"configcenter/src/common/metadata"
-	"configcenter/src/common/util"
-	"configcenter/src/common/util/table"
 	"configcenter/src/storage/dal"
 	"configcenter/src/storage/dal/redis"
 	"configcenter/src/storage/dal/types"
-	dtype "configcenter/src/storage/types"
 
+	"github.com/google/uuid"
 	"github.com/tidwall/gjson"
 	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
-	"go.mongodb.org/mongo-driver/mongo/readpref"
 	"go.mongodb.org/mongo-driver/x/mongo/driver/connstring"
 )
 
-// Mongo TODO
+// Mongo is the mongo client for one tenant
 type Mongo struct {
-	dbc    *mongo.Client
-	dbname string
-	sess   mongo.Session
-	tm     *TxnManager
-	conf   *mongoCliConf
+	// tenant is the tenant id
+	tenant string
+	// ignoreTenant is used for platform operations that do not need to specify tenant
+	ignoreTenant bool
+	*mongoClient
+	tm   *TxnManager
+	conf *mongoCliConf
+
+	// TODO: remove this when all db clients use ShardingDB
+	enableSharding bool
 }
 
 // mongoCliConf is cmdb mongo client config
@@ -61,26 +62,50 @@ type mongoCliConf struct {
 
 var _ dal.DB = new(Mongo)
 
-// MongoConf TODO
-type MongoConf struct {
-	TimeoutSeconds int
-	MaxOpenConns   uint64
-	MaxIdleConns   uint64
-	URI            string
-	RsName         string
-	SocketTimeout  int
-	DisableInsert  bool
-}
-
 // NewMgo returns new RDB
 func NewMgo(config MongoConf, timeout time.Duration) (*Mongo, error) {
+	client, err := newMongoClient(true, "", config, timeout)
+	if err != nil {
+		return nil, err
+	}
+
+	mgo := &Mongo{
+		mongoClient: client,
+		tm:          new(TxnManager),
+		conf: &mongoCliConf{
+			disableInsert: config.DisableInsert,
+		},
+	}
+
+	ctx := context.Background()
+	mgo.conf.idGenStep, err = mgo.initIDGenerator(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	return mgo, nil
+}
+
+type mongoClient struct {
+	dbc      *mongo.Client
+	dbname   string
+	uuid     string
+	disabled bool
+}
+
+// newMongoClient new mongodb client
+func newMongoClient(isMaster bool, uuid string, config MongoConf, timeout time.Duration) (*mongoClient, error) {
 	connStr, err := connstring.Parse(config.URI)
-	if nil != err {
+	if err != nil {
 		return nil, err
 	}
 	if config.RsName == "" {
 		return nil, fmt.Errorf("mongodb rsName not set")
 	}
+	if !isMaster && uuid == "" {
+		return nil, fmt.Errorf("mongodb client uuid not set")
+	}
+
 	socketTimeout := time.Second * time.Duration(config.SocketTimeout)
 	maxConnIdleTime := 25 * time.Minute
 	appName := common.GetIdentification()
@@ -111,27 +136,31 @@ func NewMgo(config MongoConf, timeout time.Duration) (*Mongo, error) {
 	// initialize mongodb related metrics
 	initMongoMetric()
 
-	mgo := &Mongo{
-		dbc:    client,
-		dbname: connStr.Database,
-		tm:     &TxnManager{},
-		conf: &mongoCliConf{
-			disableInsert: config.DisableInsert,
-		},
+	return &mongoClient{
+		dbc:      client,
+		dbname:   connStr.Database,
+		uuid:     uuid,
+		disabled: config.Disabled,
+	}, nil
+}
+
+// convColl convert collection name by tenant
+func (m *mongoClient) convColl(ignoreTenant bool, tenant, collection string) (string, error) {
+	if ignoreTenant {
+		if !common.IsPlatformTable(collection) {
+			return "", fmt.Errorf("tenant table %s has no tenant", collection)
+		}
+		return collection, nil
 	}
 
-	mgo.conf.idGenStep, err = mgo.initIDGenerator()
-	if err != nil {
-		return nil, err
+	if common.IsPlatformTable(collection) {
+		return "", fmt.Errorf("platform table %s do not need tenant", collection)
 	}
-
-	return mgo, nil
+	return common.GenTenantTableName(tenant, collection), nil
 }
 
 // initIDGenerator init id generator by config admin, returns id generator step
-func (c *Mongo) initIDGenerator() (int, error) {
-	ctx := context.Background()
-
+func (c *Mongo) initIDGenerator(ctx context.Context) (int, error) {
 	cond := map[string]interface{}{"_id": common.ConfigAdminID}
 
 	confData := make(map[string]string)
@@ -227,7 +256,7 @@ func (c *Mongo) updateIDGenSeqID(ctx context.Context, typ idgen.IDGenType, id ui
 }
 
 // checkMongodbVersion TODO
-// from now on, mongodb version must >= 4.2.0
+// from now on, mongodb version must >= 7.0
 func checkMongodbVersion(db string, client *mongo.Client) error {
 	serverStatus, err := client.Database(db).RunCommand(
 		context.Background(),
@@ -246,26 +275,48 @@ func checkMongodbVersion(db string, client *mongo.Client) error {
 	if len(fields) != 3 {
 		return fmt.Errorf("got invalid mongodb version: %v", version.StringValue())
 	}
-	// version must be >= v4.2.0
+	// version must be >= v7.0
 	major, err := strconv.Atoi(fields[0])
 	if err != nil {
 		return fmt.Errorf("parse mongodb version %s major failed, err: %v", version.StringValue(), err)
 	}
-	if major < 4 {
-		return errors.New("mongodb version must be >= v4.2.0")
+	if major < 7 {
+		return errors.New("mongodb version must be >= v7.0")
 	}
 
-	minor, err := strconv.Atoi(fields[1])
-	if err != nil {
-		return fmt.Errorf("parse mongodb version %s minor failed, err: %v", version.StringValue(), err)
-	}
-	if minor < 2 {
-		return errors.New("mongodb version must be >= v4.2.0")
-	}
 	return nil
 }
 
+// getShardingDBConfig get sharding db config
+func (c *Mongo) getShardingDBConfig(ctx context.Context) (*ShardingDBConf, error) {
+	cond := map[string]any{common.MongoMetaID: common.ShardingDBConfID}
+	conf := new(ShardingDBConf)
+	err := c.Table(common.BKTableNameSystem).Find(cond).One(ctx, &conf)
+	if err != nil {
+		if !c.IsNotFoundError(err) {
+			return nil, fmt.Errorf("get sharding db config failed, err: %v", err)
+		}
+
+		// generate new sharding db config and save it if not exists, new tenant will be added to master db by default
+		newUUID := uuid.NewString()
+		conf = &ShardingDBConf{
+			ID:             common.ShardingDBConfID,
+			MasterDB:       newUUID,
+			AddNewTenantDB: newUUID,
+			SlaveDB:        make(map[string]MongoConf),
+		}
+		if err = c.Table(common.BKTableNameSystem).Insert(ctx, conf); err != nil {
+			return nil, fmt.Errorf("insert new sharding db config failed, err: %v", err)
+		}
+		return conf, nil
+	}
+	blog.Infof("slave mongo config: %+v", conf.SlaveDB)
+
+	return conf, nil
+}
+
 // InitTxnManager TxnID management of initial transaction
+// TODO remove this
 func (c *Mongo) InitTxnManager(r redis.Client) error {
 	return c.tm.InitTxnManager(r)
 }
@@ -313,6 +364,14 @@ func (c *Mongo) IsNotFoundError(err error) bool {
 
 // Table collection operation
 func (c *Mongo) Table(collName string) types.Table {
+	if c.enableSharding {
+		var err error
+		collName, err = c.mongoClient.convColl(c.ignoreTenant, c.tenant, collName)
+		if err != nil {
+			return &errColl{err: err}
+		}
+	}
+
 	col := Collection{}
 	col.collName = collName
 	col.Mongo = c
@@ -331,483 +390,6 @@ func (c *Mongo) GetDBName() string {
 	return c.dbname
 }
 
-// Collection implement client.Collection interface
-type Collection struct {
-	collName string // 集合名
-	*Mongo
-}
-
-// Find 查询多个并反序列化到 Result
-func (c *Collection) Find(filter types.Filter, opts ...*types.FindOpts) types.Find {
-	find := &Find{
-		Collection: c,
-		filter:     filter,
-		projection: make(map[string]int),
-	}
-
-	find.Option(opts...)
-
-	return find
-}
-
-// Find define a find operation
-type Find struct {
-	*Collection
-
-	projection map[string]int
-	filter     types.Filter
-	start      int64
-	limit      int64
-	sort       bson.D
-
-	option types.FindOpts
-}
-
-// Fields 查询字段
-func (f *Find) Fields(fields ...string) types.Find {
-	for _, field := range fields {
-		if len(field) <= 0 {
-			continue
-		}
-		f.projection[field] = 1
-	}
-	return f
-}
-
-// Sort 查询排序
-// sort支持多字段最左原则排序
-// sort值为"host_id, -host_name"和sort值为"host_id:1, host_name:-1"是一样的，都代表先按host_id递增排序，再按host_name递减排序
-func (f *Find) Sort(sort string) types.Find {
-	if sort != "" {
-		sortArr := strings.Split(sort, ",")
-		f.sort = bson.D{}
-		for _, sortItem := range sortArr {
-			sortItemArr := strings.Split(strings.TrimSpace(sortItem), ":")
-			sortKey := strings.TrimLeft(sortItemArr[0], "+-")
-			if len(sortItemArr) == 2 {
-				sortDescFlag := strings.TrimSpace(sortItemArr[1])
-				if sortDescFlag == "-1" {
-					f.sort = append(f.sort, bson.E{sortKey, -1})
-				} else {
-					f.sort = append(f.sort, bson.E{sortKey, 1})
-				}
-			} else {
-				if strings.HasPrefix(sortItemArr[0], "-") {
-					f.sort = append(f.sort, bson.E{sortKey, -1})
-				} else {
-					f.sort = append(f.sort, bson.E{sortKey, 1})
-				}
-			}
-		}
-	}
-
-	return f
-}
-
-// Start 查询上标
-func (f *Find) Start(start uint64) types.Find {
-	// change to int64,后续改成int64
-	dbStart := int64(start)
-	f.start = dbStart
-	return f
-}
-
-// Limit 查询限制
-func (f *Find) Limit(limit uint64) types.Find {
-	// change to int64,后续改成int64
-	dbLimit := int64(limit)
-	f.limit = dbLimit
-	return f
-}
-
-var hostSpecialFieldMap = map[string]bool{
-	common.BKHostInnerIPField:   true,
-	common.BKHostOuterIPField:   true,
-	common.BKOperatorField:      true,
-	common.BKBakOperatorField:   true,
-	common.BKHostInnerIPv6Field: true,
-	common.BKHostOuterIPv6Field: true,
-}
-
-// All 查询多个
-func (f *Find) All(ctx context.Context, result interface{}) error {
-	mtc.collectOperCount(f.collName, findOper)
-
-	rid := ctx.Value(common.ContextRequestIDField)
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(f.collName, findOper, time.Since(start))
-	}()
-
-	err := validHostType(f.collName, f.projection, result, rid)
-	if err != nil {
-		return err
-	}
-
-	findOpts := f.generateMongoOption()
-	// 查询条件为空时候，mongodb 不返回数据
-	if f.filter == nil {
-		f.filter = bson.M{}
-	}
-
-	opt := getCollectionOption(ctx)
-
-	return f.tm.AutoRunWithTxn(ctx, f.dbc, func(ctx context.Context) error {
-		cursor, err := f.dbc.Database(f.dbname).Collection(f.collName, opt).Find(ctx, f.filter, findOpts)
-		if err != nil {
-			mtc.collectErrorCount(f.collName, findOper)
-			return err
-		}
-		return cursor.All(ctx, result)
-	})
-}
-
-// List 查询多个数据， 当分页中start值为零的时候返回满足条件总行数
-func (f *Find) List(ctx context.Context, result interface{}) (int64, error) {
-	mtc.collectOperCount(f.collName, findOper)
-
-	rid := ctx.Value(common.ContextRequestIDField)
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(f.collName, findOper, time.Since(start))
-	}()
-
-	err := validHostType(f.collName, f.projection, result, rid)
-	if err != nil {
-		return 0, err
-	}
-
-	findOpts := f.generateMongoOption()
-	// 查询条件为空时候，mongodb 不返回数据
-	if f.filter == nil {
-		f.filter = bson.M{}
-	}
-
-	opt := getCollectionOption(ctx)
-
-	var total int64
-	err = f.tm.AutoRunWithTxn(ctx, f.dbc, func(ctx context.Context) error {
-		if f.start == 0 || (f.option.WithCount != nil && *f.option.WithCount) {
-			var cntErr error
-			total, cntErr = f.dbc.Database(f.dbname).Collection(f.collName, opt).CountDocuments(ctx, f.filter)
-			if cntErr != nil {
-				return cntErr
-			}
-		}
-		cursor, err := f.dbc.Database(f.dbname).Collection(f.collName, opt).Find(ctx, f.filter, findOpts)
-		if err != nil {
-			mtc.collectErrorCount(f.collName, findOper)
-			return err
-		}
-		return cursor.All(ctx, result)
-	})
-
-	return total, nil
-}
-
-// One 查询一个
-func (f *Find) One(ctx context.Context, result interface{}) error {
-	mtc.collectOperCount(f.collName, findOper)
-
-	start := time.Now()
-	rid := ctx.Value(common.ContextRequestIDField)
-	defer func() {
-		mtc.collectOperDuration(f.collName, findOper, time.Since(start))
-	}()
-
-	err := validHostType(f.collName, f.projection, result, rid)
-	if err != nil {
-		return err
-	}
-
-	findOpts := f.generateMongoOption()
-
-	// 查询条件为空时候，mongodb panic
-	if f.filter == nil {
-		f.filter = bson.M{}
-	}
-
-	opt := getCollectionOption(ctx)
-	return f.tm.AutoRunWithTxn(ctx, f.dbc, func(ctx context.Context) error {
-		cursor, err := f.dbc.Database(f.dbname).Collection(f.collName, opt).Find(ctx, f.filter, findOpts)
-		if err != nil {
-			mtc.collectErrorCount(f.collName, findOper)
-			return err
-		}
-
-		defer cursor.Close(ctx)
-		for cursor.Next(ctx) {
-			return cursor.Decode(result)
-		}
-		return types.ErrDocumentNotFound
-	})
-
-}
-
-// Count 统计数量(非事务)
-func (f *Find) Count(ctx context.Context) (uint64, error) {
-	mtc.collectOperCount(f.collName, countOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(f.collName, countOper, time.Since(start))
-	}()
-
-	if f.filter == nil {
-		f.filter = bson.M{}
-	}
-
-	opt := getCollectionOption(ctx)
-
-	sessCtx, _, useTxn, err := f.tm.GetTxnContext(ctx, f.dbc)
-	if err != nil {
-		return 0, err
-	}
-	if !useTxn {
-		// not use transaction.
-		cnt, err := f.dbc.Database(f.dbname).Collection(f.collName, opt).CountDocuments(ctx, f.filter)
-		if err != nil {
-			mtc.collectErrorCount(f.collName, countOper)
-			return 0, err
-		}
-
-		return uint64(cnt), err
-	} else {
-		// use transaction
-		cnt, err := f.dbc.Database(f.dbname).Collection(f.collName, opt).CountDocuments(sessCtx, f.filter)
-		// do not release th session, otherwise, the session will be returned to the
-		// session pool and will be reused. then mongodb driver will increase the transaction number
-		// automatically and do read/write retry if policy is set.
-		// mongo.CmdbReleaseSession(ctx, session)
-		if err != nil {
-			mtc.collectErrorCount(f.collName, countOper)
-			return 0, err
-		}
-		return uint64(cnt), nil
-	}
-}
-
-// Insert 插入数据, docs 可以为 单个数据 或者 多个数据
-func (c *Collection) Insert(ctx context.Context, docs interface{}) error {
-	mtc.collectOperCount(c.collName, insertOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, insertOper, time.Since(start))
-	}()
-
-	rows := util.ConvertToInterfaceSlice(docs)
-
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).InsertMany(ctx, rows)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, insertOper)
-			return err
-		}
-
-		return nil
-	})
-}
-
-// Update 更新数据
-func (c *Collection) Update(ctx context.Context, filter types.Filter, doc interface{}) error {
-	mtc.collectOperCount(c.collName, updateOper)
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, updateOper, time.Since(start))
-	}()
-
-	if filter == nil {
-		filter = bson.M{}
-	}
-
-	data := bson.M{"$set": doc}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, filter, data)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, updateOper)
-			return err
-		}
-		return nil
-	})
-}
-
-// UpdateMany TODO
-// Update 更新数据, 返回修改成功的条数
-func (c *Collection) UpdateMany(ctx context.Context, filter types.Filter, doc interface{}) (uint64, error) {
-	mtc.collectOperCount(c.collName, updateOper)
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, updateOper, time.Since(start))
-	}()
-
-	if filter == nil {
-		filter = bson.M{}
-	}
-
-	data := bson.M{"$set": doc}
-	var modifiedCount uint64
-	err := c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		updateRet, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, filter, data)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, updateOper)
-			return err
-		}
-		modifiedCount = uint64(updateRet.ModifiedCount)
-		return nil
-	})
-	return modifiedCount, err
-}
-
-// Upsert 数据存在更新数据，否则新加数据。
-// 注意：该接口非原子操作，可能存在插入多条相同数据的风险。
-func (c *Collection) Upsert(ctx context.Context, filter types.Filter, doc interface{}) error {
-	mtc.collectOperCount(c.collName, upsertOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, upsertOper, time.Since(start))
-	}()
-
-	// set upsert option
-	doUpsert := true
-	replaceOpt := &options.UpdateOptions{
-		Upsert: &doUpsert,
-	}
-	data := bson.M{"$set": doc}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateOne(ctx, filter, data, replaceOpt)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, upsertOper)
-			return err
-		}
-		return nil
-	})
-
-}
-
-// UpdateMultiModel 根据不同的操作符去更新数据
-func (c *Collection) UpdateMultiModel(ctx context.Context, filter types.Filter, updateModel ...types.ModeUpdate) error {
-	mtc.collectOperCount(c.collName, updateOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, updateOper, time.Since(start))
-	}()
-
-	data := bson.M{}
-	for _, item := range updateModel {
-		if _, ok := data[item.Op]; ok {
-			return errors.New(item.Op + " appear multiple times")
-		}
-		data["$"+item.Op] = item.Doc
-	}
-
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, filter, data)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, updateOper)
-			return err
-		}
-		return nil
-	})
-
-}
-
-// Delete 删除数据
-func (c *Collection) Delete(ctx context.Context, filter types.Filter) error {
-	_, err := c.DeleteMany(ctx, filter)
-	return err
-}
-
-// DeleteMany TODO
-// Delete 删除数据， 返回删除的行数
-func (c *Collection) DeleteMany(ctx context.Context, filter types.Filter) (uint64, error) {
-	mtc.collectOperCount(c.collName, deleteOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, deleteOper, time.Since(start))
-	}()
-
-	var deleteCount uint64
-	err := c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		if err := c.tryArchiveDeletedDoc(ctx, filter); err != nil {
-			mtc.collectErrorCount(c.collName, deleteOper)
-			return err
-		}
-		deleteRet, err := c.dbc.Database(c.dbname).Collection(c.collName).DeleteMany(ctx, filter)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, deleteOper)
-			return err
-		}
-
-		deleteCount = uint64(deleteRet.DeletedCount)
-		return nil
-	})
-
-	return deleteCount, err
-}
-
-func (c *Collection) tryArchiveDeletedDoc(ctx context.Context, filter types.Filter) error {
-	delArchiveTable, exists := table.GetDelArchiveTable(c.collName)
-	if !exists {
-		// do not archive the delete docs
-		return nil
-	}
-
-	// only archive the specified fields for delete docs
-	var findOpts *options.FindOptions
-	fields := table.GetDelArchiveFields(c.collName)
-	if len(fields) > 0 {
-		projection := map[string]int{"_id": 1}
-		for _, field := range fields {
-			projection[field] = 1
-		}
-		findOpts = &options.FindOptions{Projection: projection}
-	}
-
-	docs := make([]bson.D, 0)
-	cursor, err := c.dbc.Database(c.dbname).Collection(c.collName).Find(ctx, filter, findOpts)
-	if err != nil {
-		return err
-	}
-
-	if err := cursor.All(ctx, &docs); err != nil {
-		return err
-	}
-
-	if len(docs) == 0 {
-		return nil
-	}
-
-	archives := make([]interface{}, len(docs))
-	for idx, doc := range docs {
-		detail := make(bson.D, 0)
-		var oid string
-		for _, e := range doc {
-			if e.Key == "_id" {
-				rawOid, ok := e.Value.(primitive.ObjectID)
-				if !ok {
-					return errors.New("invalid object id")
-				}
-				oid = rawOid.Hex()
-				continue
-			}
-			detail = append(detail, e)
-		}
-		archives[idx] = metadata.DeleteArchive{
-			Oid:    oid,
-			Detail: detail,
-			Time:   time.Now(),
-			Coll:   c.collName,
-		}
-	}
-
-	_, err = c.dbc.Database(c.dbname).Collection(delArchiveTable).InsertMany(ctx, archives)
-	return err
-}
-
 func (c *Mongo) redirectTable(tableName string) string {
 	if common.IsObjectInstShardingTable(tableName) {
 		tableName = common.BKTableNameBaseInst
@@ -819,6 +401,10 @@ func (c *Mongo) redirectTable(tableName string) string {
 
 // NextSequence 获取新序列号(非事务)
 func (c *Mongo) NextSequence(ctx context.Context, sequenceName string) (uint64, error) {
+	if c.enableSharding && !c.ignoreTenant {
+		return 0, errors.New("next sequence do not need tenant")
+	}
+
 	sequenceName = c.redirectTable(sequenceName)
 
 	rid := ctx.Value(common.ContextRequestIDField)
@@ -855,6 +441,10 @@ func (c *Mongo) NextSequence(ctx context.Context, sequenceName string) (uint64, 
 
 // NextSequences 批量获取新序列号(非事务)
 func (c *Mongo) NextSequences(ctx context.Context, sequenceName string, num int) ([]uint64, error) {
+	if c.enableSharding && !c.ignoreTenant {
+		return nil, errors.New("next sequence do not need tenant")
+	}
+
 	if num == 0 {
 		return make([]uint64, 0), nil
 	}
@@ -910,6 +500,14 @@ type Idgen struct {
 
 // HasTable 判断是否存在集合
 func (c *Mongo) HasTable(ctx context.Context, collName string) (bool, error) {
+	if c.enableSharding {
+		var err error
+		collName, err = c.mongoClient.convColl(c.ignoreTenant, c.tenant, collName)
+		if err != nil {
+			return false, err
+		}
+	}
+
 	cursor, err := c.dbc.Database(c.dbname).ListCollections(ctx, bson.M{"name": collName, "type": "collection"})
 	if err != nil {
 		return false, err
@@ -925,563 +523,58 @@ func (c *Mongo) HasTable(ctx context.Context, collName string) (bool, error) {
 
 // ListTables 获取所有的表名
 func (c *Mongo) ListTables(ctx context.Context) ([]string, error) {
-	return c.dbc.Database(c.dbname).ListCollectionNames(ctx, bson.M{"type": "collection"})
-
+	filter := bson.M{"type": "collection"}
+	if c.enableSharding {
+		if c.ignoreTenant {
+			filter["name"] = bson.M{common.BKDBIN: common.PlatformTables()}
+		} else {
+			filter["name"] = bson.M{common.BKDBLIKE: fmt.Sprintf("^%s_", c.tenant)}
+		}
+	}
+	return c.dbc.Database(c.dbname).ListCollectionNames(ctx, filter)
 }
 
 // DropTable 移除集合
 func (c *Mongo) DropTable(ctx context.Context, collName string) error {
+	if c.enableSharding {
+		var err error
+		collName, err = c.mongoClient.convColl(c.ignoreTenant, c.tenant, collName)
+		if err != nil {
+			return err
+		}
+	}
 	return c.dbc.Database(c.dbname).Collection(collName).Drop(ctx)
 }
 
 // CreateTable 创建集合 TODO test
 func (c *Mongo) CreateTable(ctx context.Context, collName string) error {
+	if c.enableSharding {
+		var err error
+		collName, err = c.mongoClient.convColl(c.ignoreTenant, c.tenant, collName)
+		if err != nil {
+			return err
+		}
+	}
 	return c.dbc.Database(c.dbname).RunCommand(ctx, map[string]interface{}{"create": collName}).Err()
 }
 
 // RenameTable 更新集合名称
 func (c *Mongo) RenameTable(ctx context.Context, prevName, currName string) error {
+	if c.enableSharding {
+		var err error
+		prevName, err = c.mongoClient.convColl(c.ignoreTenant, c.tenant, prevName)
+		if err != nil {
+			return err
+		}
+		currName, err = c.mongoClient.convColl(c.ignoreTenant, c.tenant, currName)
+		if err != nil {
+			return err
+		}
+	}
+
 	cmd := bson.D{
 		{"renameCollection", c.dbname + "." + prevName},
 		{"to", c.dbname + "." + currName},
 	}
 	return c.dbc.Database("admin").RunCommand(ctx, cmd).Err()
-}
-
-// BatchCreateIndexes 批量创建索引
-func (c *Collection) BatchCreateIndexes(ctx context.Context, indexes []types.Index) error {
-	mtc.collectOperCount(c.collName, indexCreateOper)
-
-	createIndexInfos := make([]mongo.IndexModel, len(indexes))
-	for idx, index := range indexes {
-		createIndexInfo, err := buildIndex(index)
-		if err != nil {
-			return err
-		}
-
-		createIndexInfos[idx] = createIndexInfo
-	}
-
-	indexView := c.dbc.Database(c.dbname).Collection(c.collName).Indexes()
-	_, err := indexView.CreateMany(ctx, createIndexInfos)
-	if err != nil {
-		mtc.collectErrorCount(c.collName, indexCreateOper)
-		// ignore the following case
-		// 1.the new index is exactly the same as the existing one
-		// 2.the new index has same keys with the existing one, but its name is different
-		if strings.Contains(err.Error(), "all indexes already exist") ||
-			strings.Contains(err.Error(), "already exists with a different name") {
-			return nil
-		}
-	}
-
-	return err
-}
-
-// CreateIndex 创建索引
-func (c *Collection) CreateIndex(ctx context.Context, index types.Index) error {
-	mtc.collectOperCount(c.collName, indexCreateOper)
-
-	createIndexInfo, err := buildIndex(index)
-	if err != nil {
-		return err
-	}
-
-	indexView := c.dbc.Database(c.dbname).Collection(c.collName).Indexes()
-	_, err = indexView.CreateOne(ctx, createIndexInfo)
-	if err != nil {
-		mtc.collectErrorCount(c.collName, indexCreateOper)
-		// ignore the following case
-		// 1.the new index is exactly the same as the existing one
-		// 2.the new index has same keys with the existing one, but its name is different
-		if strings.Contains(err.Error(), "all indexes already exist") ||
-			strings.Contains(err.Error(), "already exists with a different name") {
-			return nil
-		}
-	}
-
-	return err
-}
-
-func buildIndex(index types.Index) (mongo.IndexModel, error) {
-	createIndexOpt := &options.IndexOptions{
-		Background:              &index.Background,
-		Unique:                  &index.Unique,
-		PartialFilterExpression: index.PartialFilterExpression,
-	}
-	if index.Name != "" {
-		createIndexOpt.Name = &index.Name
-	}
-
-	if index.ExpireAfterSeconds != 0 {
-		createIndexOpt.SetExpireAfterSeconds(index.ExpireAfterSeconds)
-	}
-
-	keys := index.Keys
-	for idx, key := range keys {
-		val, err := util.GetInt32ByInterface(key.Value)
-		if err != nil {
-			return mongo.IndexModel{}, err
-		}
-		key.Value = val
-		keys[idx] = key
-	}
-
-	return mongo.IndexModel{
-		Keys:    keys,
-		Options: createIndexOpt,
-	}, nil
-}
-
-// DropIndex remove index by name
-func (c *Collection) DropIndex(ctx context.Context, indexName string) error {
-	mtc.collectOperCount(c.collName, indexDropOper)
-	indexView := c.dbc.Database(c.dbname).Collection(c.collName).Indexes()
-	_, err := indexView.DropOne(ctx, indexName)
-	if err != nil {
-		if strings.Contains(err.Error(), "IndexNotFound") {
-			return nil
-		}
-		mtc.collectErrorCount(c.collName, indexDropOper)
-		return err
-	}
-	return nil
-}
-
-// Indexes get all indexes for the collection
-func (c *Collection) Indexes(ctx context.Context) ([]types.Index, error) {
-	indexView := c.dbc.Database(c.dbname).Collection(c.collName).Indexes()
-	cursor, err := indexView.List(ctx)
-	if nil != err {
-		return nil, err
-	}
-	defer cursor.Close(ctx)
-	var indexes []types.Index
-	for cursor.Next(ctx) {
-		idxResult := types.Index{}
-		cursor.Decode(&idxResult)
-		indexes = append(indexes, idxResult)
-	}
-
-	return indexes, nil
-}
-
-// AddColumn add a new column for the collection
-func (c *Collection) AddColumn(ctx context.Context, column string, value interface{}) error {
-	mtc.collectOperCount(c.collName, columnOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, columnOper, time.Since(start))
-	}()
-
-	selector := dtype.Document{column: dtype.Document{"$exists": false}}
-	datac := dtype.Document{"$set": dtype.Document{column: value}}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, selector, datac)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, columnOper)
-			return err
-		}
-		return nil
-	})
-}
-
-// RenameColumn rename a column for the collection
-func (c *Collection) RenameColumn(ctx context.Context, filter types.Filter, oldName, newColumn string) error {
-	mtc.collectOperCount(c.collName, columnOper)
-	if filter == nil {
-		filter = dtype.Document{}
-	}
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, columnOper, time.Since(start))
-	}()
-
-	datac := dtype.Document{"$rename": dtype.Document{oldName: newColumn}}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, filter, datac)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, columnOper)
-			return err
-		}
-
-		return nil
-	})
-}
-
-// DropColumn remove a column by the name
-func (c *Collection) DropColumn(ctx context.Context, field string) error {
-	mtc.collectOperCount(c.collName, columnOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, columnOper, time.Since(start))
-	}()
-
-	datac := dtype.Document{"$unset": dtype.Document{field: ""}}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, dtype.Document{}, datac)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, columnOper)
-			return err
-		}
-
-		return nil
-	})
-}
-
-// DropColumns remove many columns by the name
-func (c *Collection) DropColumns(ctx context.Context, filter types.Filter, fields []string) error {
-	mtc.collectOperCount(c.collName, columnOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, columnOper, time.Since(start))
-	}()
-
-	unsetFields := make(map[string]interface{})
-	for _, field := range fields {
-		unsetFields[field] = ""
-	}
-
-	datac := dtype.Document{"$unset": unsetFields}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, filter, datac)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, columnOper)
-			return err
-		}
-
-		return nil
-	})
-}
-
-// DropDocsColumn remove a column by the name for doc use filter
-func (c *Collection) DropDocsColumn(ctx context.Context, field string, filter types.Filter) error {
-	mtc.collectOperCount(c.collName, columnOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, columnOper, time.Since(start))
-	}()
-
-	// 查询条件为空时候，mongodb 不返回数据
-	if filter == nil {
-		filter = bson.M{}
-	}
-
-	datac := dtype.Document{"$unset": dtype.Document{field: ""}}
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		_, err := c.dbc.Database(c.dbname).Collection(c.collName).UpdateMany(ctx, filter, datac)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, columnOper)
-			return err
-		}
-
-		return nil
-	})
-}
-
-// AggregateAll aggregate all operation
-func (c *Collection) AggregateAll(ctx context.Context, pipeline interface{}, result interface{},
-	opts ...*types.AggregateOpts) error {
-
-	mtc.collectOperCount(c.collName, aggregateOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, aggregateOper, time.Since(start))
-	}()
-
-	var aggregateOption *options.AggregateOptions
-	for _, opt := range opts {
-		if opt == nil {
-			continue
-		}
-		if opt.AllowDiskUse != nil {
-			aggregateOption = &options.AggregateOptions{AllowDiskUse: opt.AllowDiskUse}
-		}
-	}
-
-	opt := getCollectionOption(ctx)
-
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		cursor, err := c.dbc.Database(c.dbname).Collection(c.collName, opt).Aggregate(ctx, pipeline, aggregateOption)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, aggregateOper)
-			return err
-		}
-		defer cursor.Close(ctx)
-		return decodeCursorIntoSlice(ctx, cursor, result)
-	})
-
-}
-
-// AggregateOne aggregate one operation
-func (c *Collection) AggregateOne(ctx context.Context, pipeline interface{}, result interface{}) error {
-	mtc.collectOperCount(c.collName, aggregateOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, aggregateOper, time.Since(start))
-	}()
-
-	opt := getCollectionOption(ctx)
-
-	return c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		cursor, err := c.dbc.Database(c.dbname).Collection(c.collName, opt).Aggregate(ctx, pipeline)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, aggregateOper)
-			return err
-		}
-
-		defer cursor.Close(ctx)
-		for cursor.Next(ctx) {
-			return cursor.Decode(result)
-		}
-		return types.ErrDocumentNotFound
-	})
-
-}
-
-// Distinct Finds the distinct values for a specified field across a single collection or view and returns the results in an
-// field the field for which to return distinct values.
-// filter query that specifies the documents from which to retrieve the distinct values.
-func (c *Collection) Distinct(ctx context.Context, field string, filter types.Filter) ([]interface{}, error) {
-	mtc.collectOperCount(c.collName, distinctOper)
-
-	start := time.Now()
-	defer func() {
-		mtc.collectOperDuration(c.collName, distinctOper, time.Since(start))
-	}()
-
-	if filter == nil {
-		filter = bson.M{}
-	}
-
-	opt := getCollectionOption(ctx)
-	var results []interface{} = nil
-	err := c.tm.AutoRunWithTxn(ctx, c.dbc, func(ctx context.Context) error {
-		var err error
-		results, err = c.dbc.Database(c.dbname).Collection(c.collName, opt).Distinct(ctx, field, filter)
-		if err != nil {
-			mtc.collectErrorCount(c.collName, distinctOper)
-			return err
-		}
-
-		return nil
-	})
-	return results, err
-}
-
-// Option TODO
-func (f *Find) Option(opts ...*types.FindOpts) {
-	for _, opt := range opts {
-		if opt == nil {
-			continue
-		}
-		if opt.WithObjectID != nil {
-			f.option.WithObjectID = opt.WithObjectID
-		}
-		if opt.WithCount != nil {
-			f.option.WithCount = opt.WithCount
-		}
-	}
-}
-
-func (f *Find) generateMongoOption() *options.FindOptions {
-	findOpts := &options.FindOptions{}
-	if f.projection == nil {
-		f.projection = make(map[string]int, 0)
-	}
-	if f.option.WithObjectID != nil && *f.option.WithObjectID {
-		// mongodb 要求，当有字段设置未1, 不设置都不显示
-		// 没有设置projection 的时候，返回所有字段
-		if len(f.projection) > 0 {
-			f.projection["_id"] = 1
-		}
-	} else {
-		if _, exists := f.projection["_id"]; !exists {
-			f.projection["_id"] = 0
-		}
-	}
-	if len(f.projection) != 0 {
-		findOpts.Projection = f.projection
-	}
-
-	if f.start != 0 {
-		findOpts.SetSkip(f.start)
-	}
-	if f.limit != 0 {
-		findOpts.SetLimit(f.limit)
-	}
-	if len(f.sort) != 0 {
-		findOpts.SetSort(f.sort)
-	}
-
-	return findOpts
-}
-
-func decodeCursorIntoSlice(ctx context.Context, cursor *mongo.Cursor, result interface{}) error {
-	resultv := reflect.ValueOf(result)
-	if resultv.Kind() != reflect.Ptr || resultv.Elem().Kind() != reflect.Slice {
-		return errors.New("result argument must be a slice address")
-	}
-
-	elemt := resultv.Elem().Type().Elem()
-	slice := reflect.MakeSlice(resultv.Elem().Type(), 0, 10)
-	for cursor.Next(ctx) {
-		elemp := reflect.New(elemt)
-		if err := cursor.Decode(elemp.Interface()); nil != err {
-			return err
-		}
-		slice = reflect.Append(slice, elemp.Elem())
-	}
-	if err := cursor.Err(); err != nil {
-		return err
-	}
-
-	resultv.Elem().Set(slice)
-	return nil
-}
-
-// validHostType valid if host query uses specified type that transforms ip & operator array to string
-func validHostType(collection string, projection map[string]int, result interface{}, rid interface{}) error {
-	if result == nil {
-		blog.Errorf("host query result is nil, rid: %s", rid)
-		return fmt.Errorf("host query result type invalid")
-	}
-
-	if collection != common.BKTableNameBaseHost {
-		return nil
-	}
-
-	// check if specified fields include special fields
-	if len(projection) != 0 {
-		needCheck := false
-		for field := range projection {
-			if hostSpecialFieldMap[field] {
-				needCheck = true
-				break
-			}
-		}
-		if !needCheck {
-			return nil
-		}
-	}
-
-	resType := reflect.TypeOf(result)
-	if resType.Kind() != reflect.Ptr {
-		blog.Errorf("host query result type(%v) not pointer type, rid: %v", resType, rid)
-		return fmt.Errorf("host query result type invalid")
-	}
-	// if result is *map[string]interface{} type, it must be *metadata.HostMapStr type
-	if resType.ConvertibleTo(reflect.TypeOf(&map[string]interface{}{})) {
-		if resType != reflect.TypeOf(&metadata.HostMapStr{}) {
-			blog.Errorf("host query result type(%v) not match *metadata.HostMapStr type, rid: %v", resType, rid)
-			return fmt.Errorf("host query result type invalid")
-		}
-		return nil
-	}
-
-	resElem := resType.Elem()
-	switch resElem.Kind() {
-	case reflect.Struct:
-		// if result is *struct type, the special field in it must be metadata.StringArrayToString type
-		numField := resElem.NumField()
-		validType := reflect.TypeOf(metadata.StringArrayToString(""))
-		for i := 0; i < numField; i++ {
-			field := resElem.Field(i)
-			bsonTag := field.Tag.Get("bson")
-			if bsonTag == "" {
-				blog.Errorf("host query result field(%s) has empty bson tag, rid: %v", field.Name, rid)
-				return fmt.Errorf("host query result type invalid")
-			}
-			if hostSpecialFieldMap[bsonTag] && field.Type != validType {
-				blog.Errorf("host query result field type(%v) not match *metadata.StringArrayToString type", field.Type)
-				return fmt.Errorf("host query result type invalid")
-			}
-		}
-	case reflect.Slice:
-		// check if slice item is valid type, map or struct validation is similar as before
-		elem := resElem.Elem()
-		if elem.ConvertibleTo(reflect.TypeOf(map[string]interface{}{})) {
-			if elem != reflect.TypeOf(metadata.HostMapStr{}) {
-				blog.Errorf("host query result type(%v) not match *[]metadata.HostMapStr type", resType)
-				return fmt.Errorf("host query result type invalid")
-			}
-			return nil
-		}
-
-		if elem.Kind() == reflect.Ptr {
-			elem = elem.Elem()
-		}
-
-		if elem.Kind() != reflect.Struct {
-			blog.Errorf("host query result type(%v) not struct pointer type or map type", resType)
-			return fmt.Errorf("host query result type invalid")
-		}
-		numField := elem.NumField()
-		validType := reflect.TypeOf(metadata.StringArrayToString(""))
-		for i := 0; i < numField; i++ {
-			field := elem.Field(i)
-			bsonTag := field.Tag.Get("bson")
-			if bsonTag == "" {
-				blog.Errorf("host query result field(%s) has empty bson tag, rid: %v", field.Name, rid)
-				return fmt.Errorf("host query result type invalid")
-			}
-			if hostSpecialFieldMap[bsonTag] && field.Type != validType {
-				blog.Errorf("host query result field type(%v) not match *metadata.StringArrayToString type", field.Type)
-				return fmt.Errorf("host query result type invalid")
-			}
-		}
-	default:
-		blog.Errorf("host query result type(%v) not pointer of map, struct or slice, rid: %v", resType, rid)
-		return fmt.Errorf("host query result type invalid")
-	}
-	return nil
-}
-
-const (
-	// reference doc:
-	// https://docs.mongodb.com/manual/core/read-preference-staleness/#replica-set-read-preference-max-staleness
-	// this is the minimum value of maxStalenessSeconds allowed.
-	// specifying a smaller maxStalenessSeconds value will raise an error. Clients estimate secondaries’ staleness
-	// by periodically checking the latest write date of each replica set member. Since these checks are infrequent,
-	// the staleness estimate is coarse. Thus, clients cannot enforce a maxStalenessSeconds value of less than
-	// 90 seconds.
-	maxStalenessSeconds = 90 * time.Second
-)
-
-func getCollectionOption(ctx context.Context) *options.CollectionOptions {
-	var opt *options.CollectionOptions
-	switch util.GetDBReadPreference(ctx) {
-
-	case common.NilMode:
-
-	case common.PrimaryMode:
-		opt = &options.CollectionOptions{
-			ReadPreference: readpref.Primary(),
-		}
-	case common.PrimaryPreferredMode:
-		opt = &options.CollectionOptions{
-			ReadPreference: readpref.PrimaryPreferred(readpref.WithMaxStaleness(maxStalenessSeconds)),
-		}
-	case common.SecondaryMode:
-		opt = &options.CollectionOptions{
-			ReadPreference: readpref.Secondary(readpref.WithMaxStaleness(maxStalenessSeconds)),
-		}
-	case common.SecondaryPreferredMode:
-		opt = &options.CollectionOptions{
-			ReadPreference: readpref.SecondaryPreferred(readpref.WithMaxStaleness(maxStalenessSeconds)),
-		}
-	case common.NearestMode:
-		opt = &options.CollectionOptions{
-			ReadPreference: readpref.Nearest(readpref.WithMaxStaleness(maxStalenessSeconds)),
-		}
-	}
-
-	return opt
 }
